requirements.txt
Django==5.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
